+++ System/EamonDD/Game/Menus/ActionMenus/UtilitiesMenus/AdventureSupportMenus/DeleteCustomAdventureClassesMenu.cs
@@ -8,6 +8,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
 using System.Text;
+using System.Text.RegularExpressions;
 using Eamon.Game.Attributes;
 using EamonDD.Framework.Menus.ActionMenus;
 using static EamonDD.Game.Plugin.PluginContext;
@@ -91,6 +92,8 @@ namespace EamonDD.Game.Menus.ActionMenus
 		/// <summary></summary>
 		protected virtual void DeleteSelectedClassFiles()
 		{
+			var regex = new Regex("^[a-zA-Z0-9]*$");
+
 			Globals.Out.Print("{0}", Globals.LineSep);
 
 			Globals.Out.WriteLine();
@@ -113,7 +116,10 @@ namespace EamonDD.Game.Menus.ActionMenus
 				}
 			}
 
-			Globals.Directory.DeleteEmptySubdirectories(@"..\" + AdventureName, true);
+			if (regex.IsMatch(AdventureName))
+			{
+				Globals.Directory.DeleteEmptySubdirectories(@"..\" + AdventureName, true);
+			}
 		}
 
 		public override void Execute()
